refactor: rename automaxprocs to prodinit
And set debug.SetTraceback("all") in there as well. The package is for initialising everything required for production.
