+++ cmd/ftl-schema/main.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/block/ftl"
 	"github.com/block/ftl/backend/schemaservice"
-	_ "github.com/block/ftl/internal/automaxprocs" // Set GOMAXPROCS to match Linux container CPU quota.
 	"github.com/block/ftl/internal/log"
 	"github.com/block/ftl/internal/observability"
+	_ "github.com/block/ftl/internal/prodinit"
 )
 
 var cli struct {
